+++ src/Http/Controllers/MediaLibraryAdminController.php
@@ -36,7 +36,7 @@ class MediaLibraryAdminController extends ModelAdminController
 
 	public function show($id)
 	{
-		$media = $this->decorator->findInstanceOrFail($id);
+		$media = $this->decorator->findInstance($id);
 
 		return Response::download(storage_path($media->getFileName()));
 	}
@@ -130,7 +130,7 @@ class MediaLibraryAdminController extends ModelAdminController
 
 	public function destroy($id)
 	{
-		$media = $this->decorator->findInstanceOrFail($id);
+		$media = $this->decorator->findInstance($id);
 		$data = [ 'files' => [
 			$media->getFilename() => true
 		]];